ml-service/pyproject.toml
[project]
name = "ml-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "easyocr>=1.7.2",
    "fastapi>=0.128.0",
    "openai>=2.14.0",
    "prophet>=1.2.1",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
    "xgboost>=3.1.2",
]
